refactor: moved mocks and test helpers to one location
